Cleans up and improves MIDI even more
